requirements.txt
flask
fpdf2
arabic-reshaper
python-bidi